semana8/requirements.txt
annotated-types==0.7.0
anyio==4.9.0
bandit==1.8.6
black==25.9.0
certifi==2025.7.14
cffi==2.0.0
cfgv==3.4.0
charset-normalizer==3.4.3
click==8.2.2
colorama==0.4.6
coverage==7.10.7
cryptography==46.0.2
distlib==0.4.0
dnspython==2.7.0
email_validator==2.2.0
factory_boy==3.3.3
Faker==37.8.0
fastapi==0.116.1
fastapi-cli==0.0.8
fastapi-cloud-cli==0.1.5
filelock==3.19.1
flake8==7.3.0
greenlet==3.2.4
h11==0.16.0
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
identify==2.6.15
idna==3.10
iniconfig==2.1.0
isort==6.1.0
itsdangerous==2.2.0
Jinja2==3.1.6
pyjwt
line_profiler==5.0.0
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mccabe==0.7.0
mdurl==0.1.2
memory-profiler==0.61.0
mypy==1.18.2
mypy_extensions==1.1.0
nodeenv==1.9.1
orjson==3.11.1
packaging==25.0
pathspec==0.12.1
platformdirs==4.4.0
pluggy==1.6.0
pre_commit==4.3.0
prometheus-fastapi-instrumentator==7.1.0
prometheus_client==0.23.1
psutil==7.1.0
py-cpuinfo==9.0.0
pycodestyle==2.14.0
pycparser==2.23
pydantic==2.11.7
pydantic-extra-types==2.10.5
pydantic-settings==2.10.1
pydantic_core==2.33.2
pyflakes==3.4.0
Pygments==2.19.2
pytest==8.4.2
pytest-asyncio==1.2.0
pytest-benchmark==5.1.0
pytest-cov==7.0.0
pytest-html==4.1.1
pytest-metadata==3.1.1
pytest-mock==3.15.1
python-dotenv==1.1.1
python-multipart==0.0.20
pytokens==0.1.10
PyYAML==6.0.2
redis==6.4.0
requests==2.32.5
requests-mock==1.12.1
rich==14.1.0
rich-toolkit==0.14.9
rignore==0.6.4
sentry-sdk==2.34.1
shellingham==1.5.4
sniffio==1.3.1
SQLAlchemy==2.0.43
starlette==0.47.2
stevedore==5.5.0
typer==0.16.0
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2025.2
ujson==5.10.0
urllib3==2.5.0
uvicorn==0.35.0
virtualenv==20.34.0
watchfiles==1.1.0
websockets==15.0.1
psycopg2-binary
